+++ loggo/__init__.py
@@ -2,7 +2,6 @@
 Simpler namespace
 """
 
-# flake8: noqa
-from .loggo import Loggo
+from .loggo import Loggo  # noqa: F401
 
 __version__ = "5.1.0"  # DO NOT EDIT THIS LINE MANUALLY. LET bump2version UTILITY DO IT